Removed unused files from billing
