+++ client/src/components/billing/BillingPage.js
@@ -1,29 +1,147 @@
 import React, { useState } from 'react';
-import { Elements, StripeProvider } from 'react-stripe-elements';
-import { withStyles } from '@material-ui/core/styles';
-import Grid from '@material-ui/core/Grid';
+import classNames from 'classnames';
+import PropTypes from 'prop-types';
+import AppBar from '@material-ui/core/AppBar';
 import Button from '@material-ui/core/Button';
-import Snackbar from '../views/Snackbar';
+import Card from '@material-ui/core/Card';
+import CardActions from '@material-ui/core/CardActions';
+import CardContent from '@material-ui/core/CardContent';
+import CardHeader from '@material-ui/core/CardHeader';
+import CssBaseline from '@material-ui/core/CssBaseline';
+import Grid from '@material-ui/core/Grid';
+import StarIcon from '@material-ui/icons/StarBorder';
+import Toolbar from '@material-ui/core/Toolbar';
+import Typography from '@material-ui/core/Typography';
+import { withStyles } from '@material-ui/core/styles';
 import TakeMoney from './TakeMoney';
-import Pricing from './Pricing';
+import { Elements, StripeProvider } from 'react-stripe-elements';
 
 const styles = theme => ({
-  cardWrapper: {
-    display: 'flex',
-    flexDirection: 'column',
+  '@global': {
+    body: {
+      backgroundColor: theme.palette.common.white
+    }
+  },
+  appBar: {
+    position: 'relative'
+  },
+  toolbarTitle: {
+    flex: 1
+  },
+  layout: {
+    width: 'auto',
+    marginLeft: theme.spacing.unit * 3,
+    marginRight: theme.spacing.unit * 3,
+    [theme.breakpoints.up(900 + theme.spacing.unit * 3 * 2)]: {
+      width: 900,
+      marginLeft: 'auto',
+      marginRight: 'auto'
+    }
+  },
+  heroContent: {
+    maxWidth: 600,
     margin: '0 auto',
-    border: '1px solid #777',
-    width: '30%',
-    height: '50vh',
-    position: 'relative',
-    justifyContent: 'space-between',
-    padding: '5%'
+    padding: `${theme.spacing.unit * 8}px 0 ${theme.spacing.unit * 6}px`
+  },
+  cardHeader: {
+    backgroundColor: theme.palette.grey[200]
+  },
+  cardPricing: {
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'baseline',
+    marginBottom: theme.spacing.unit * 2
+  },
+  cardActions: {
+    [theme.breakpoints.up('sm')]: {
+      paddingBottom: theme.spacing.unit * 2
+    }
+  },
+  footer: {
+    marginTop: theme.spacing.unit * 8,
+    borderTop: `1px solid ${theme.palette.divider}`,
+    padding: `${theme.spacing.unit * 6}px 0`
   }
 });
 
-const BillingPage = props => {
+const tiers = [
+  {
+    title: 'Standard',
+    price: '9.99',
+    value: 999,
+    description: [
+      'Up to 500 Refreshrs',
+      'Unlimited teachers',
+      'Fixed schedule times',
+      'Email support'
+    ],
+    buttonText: 'Sign up for free',
+    buttonVariant: 999
+  },
+  {
+    title: 'Premium',
+    value: 2999,
+    subheader: 'Most popular',
+    price: '29.99',
+    description: [
+      'Unlimited Refreshrs',
+      'Customize Refreshr schedule',
+      '24/7 Phone support',
+      '30-day free trial'
+    ],
+    buttonText: 'Get started',
+    buttonVariant: 2999
+  },
+  {
+    title: 'Custom',
+    price: '50+',
+    description: [
+      'Unlimited Refreshrs',
+      'Custom solution',
+      'Tailored to your school',
+      'In-person support'
+    ],
+    buttonText: 'Contact us',
+    buttonVariant: 'custom'
+  }
+];
+const footers = [
+  {
+    title: 'Company',
+    description: ['Team', 'History', 'Contact us', 'Locations']
+  },
+  {
+    title: 'Features',
+    description: [
+      'Cool stuff',
+      'Random feature',
+      'Team feature',
+      'Developer stuff',
+      'Another one'
+    ]
+  },
+  {
+    title: 'Resources',
+    description: [
+      'Resource',
+      'Resource name',
+      'Another resource',
+      'Final resource'
+    ]
+  },
+  {
+    title: 'Legal',
+    description: ['Privacy policy', 'Terms of use']
+  }
+];
+
+// function Pricing(props) {
+//   const { classes } = props;
+
+const Pricing = props => {
   const [subType, setSubType] = useState(999);
   const [snackbar, setSnackbar] = useState(false);
+  const { classes } = props;
 
   const openSnackbar = () => {
     setSnackbar(true);
@@ -33,48 +151,135 @@ const BillingPage = props => {
   };
 
   const handleChange = e => {
+    console.log('VALUE', e.target.value);
     setSubType(parseInt(e.target.value));
   };
 
   return (
-    <>
-      {/* <h1>BillingPage</h1>
-      <Button variant="contained" onClick={openSnackbar}>
-        Test Snackbar
-      </Button>
-      {snackbar && <Snackbar message="testing snackbar" />} */}
-      <Pricing />
-      {/* <Grid className={props.classes.cardWrapper}>
-        <Grid>
-          <form>
-            <label>
-              <input
-                onChange={handleChange}
-                type="radio"
-                name="subType"
-                value={999}
-              />
-              1 Year Subscription - $9.99
-            </label>
-            <label>
-              <input
-                onChange={handleChange}
-                type="radio"
-                name="subType"
-                value={2999}
-              />
-              1 Year Premium Subscription - $29.99
-            </label>
-          </form>
-          <StripeProvider apiKey="pk_test_6uEhds8mHz26DG95ZvUwTURp">
-            <Elements>
-              <TakeMoney subType={subType} />
-            </Elements>
-          </StripeProvider>
+    <React.Fragment>
+      <CssBaseline />
+      {/* <AppBar position="static" color="default" className={classes.appBar}>
+        <Toolbar>
+          <Typography
+            variant="h6"
+            color="inherit"
+            noWrap
+            className={classes.toolbarTitle}
+          >
+            Company name
+          </Typography>
+          <Button>Features</Button>
+          <Button>Enterprise</Button>
+          <Button>Support</Button>
+          <Button color="primary" variant="outlined">
+            Login
+          </Button>
+        </Toolbar>
+      </AppBar> */}
+      <main className={classes.layout}>
+        {/* Hero unit */}
+        <div className={classes.heroContent}>
+          <Typography
+            component="h1"
+            variant="h2"
+            align="center"
+            color="textPrimary"
+            gutterBottom
+          >
+            Pricing
+          </Typography>
+          <Typography
+            variant="h6"
+            align="center"
+            color="textSecondary"
+            component="p"
+          />
+        </div>
+        <Grid container spacing={40} alignItems="flex-end">
+          {tiers.map(tier => (
+            // Enterprise card is full width at sm breakpoint
+            <Grid
+              item
+              key={tier.title}
+              xs={12}
+              sm={tier.title === 'Premium' ? 12 : 6}
+              md={4}
+            >
+              <Card>
+                <CardHeader
+                  title={tier.title}
+                  subheader={tier.subheader}
+                  titleTypographyProps={{ align: 'center' }}
+                  subheaderTypographyProps={{ align: 'center' }}
+                  action={tier.title === 'Premium' ? <StarIcon /> : null}
+                  className={classes.cardHeader}
+                />
+                <CardContent>
+                  <div className={classes.cardPricing}>
+                    <Typography component="h2" variant="h3" color="textPrimary">
+                      ${tier.price}
+                    </Typography>
+                    <Typography variant="h6" color="textSecondary">
+                      /mo
+                    </Typography>
+                  </div>
+                  {tier.description.map(line => (
+                    <Typography variant="subtitle1" align="center" key={line}>
+                      {line}
+                    </Typography>
+                  ))}
+                </CardContent>
+                <CardActions className={classes.cardActions}>
+                  {/* <Button
+                    fullWidth
+                    variant={tier.buttonVariant}
+                    color="primary"
+                  >
+                    {tier.buttonText}
+                  </Button> */}
+                  <StripeProvider apiKey="pk_test_6uEhds8mHz26DG95ZvUwTURp">
+                    <Elements>
+                      <TakeMoney
+                        variant={tier.buttonVariant}
+                        subType={tiers.value}
+                      />
+                    </Elements>
+                  </StripeProvider>
+                  `
+                </CardActions>
+              </Card>
+            </Grid>
+          ))}
         </Grid>
-      </Grid> */}
-    </>
+      </main>
+      {/* Footer */}
+      {/* <footer className={classNames(classes.footer, classes.layout)}>
+        <Grid container spacing={32} justify="space-evenly">
+          {footers.map(footer => (
+            <Grid item xs key={footer.title}>
+              <Typography variant="h6" color="textPrimary" gutterBottom>
+                {footer.title}
+              </Typography>
+              {footer.description.map(item => (
+                <Typography
+                  key={item}
+                  variant="subtitle1"
+                  color="textSecondary"
+                >
+                  {item}
+                </Typography>
+              ))}
+            </Grid>
+          ))}
+        </Grid>
+      </footer> */}
+      {/* End footer */}
+    </React.Fragment>
   );
 };
 
-export default withStyles(styles)(BillingPage);
+Pricing.propTypes = {
+  classes: PropTypes.object.isRequired
+};
+
+export default withStyles(styles)(Pricing);
