+++ client/src/index.js
@@ -1,20 +1,19 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App'
-import {Auth} from './components'
+import App from './App';
+import { Auth } from './components';
 import * as serviceWorker from './serviceWorker';
 import { BrowserRouter as Router } from 'react-router-dom';
 require('dotenv').config();
 
-const auth = new Auth()
+const auth = new Auth();
 
 ReactDOM.render(
   <Router>
-    <App  auth={auth}/>
+    <App auth={auth} />
   </Router>,
-  document.getElementById('root'),
+  document.getElementById('root')
 );
-
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
